requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.51
pydantic==2.13.3
python-dotenv==1.0.0
google-generativeai==0.8.3
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
